Improve Vec example
