+++ rustlib/src/lista.rs
@@ -1,0 +1,62 @@
+use std::{mem, slice};
+
+#[no_mangle]
+extern "C" fn new_list(out_len: *mut usize) -> *const i32 {
+    let lista: Vec<i32> = Vec::new();
+    let len = lista.len();
+
+    unsafe {
+        *out_len = len;
+    }
+
+    let ptr = lista.as_ptr();
+    mem::forget(lista);
+    ptr
+}
+
+#[no_mangle]
+extern "C" fn add_item(list_ptr: *mut i32, len: usize, item: i32) -> *const i32 {
+    let mut vec = unsafe { Vec::from_raw_parts(list_ptr, len, len) };
+    vec.push(item);
+    let new_ptr = vec.as_ptr();
+    mem::forget(vec);
+    new_ptr
+}
+
+#[no_mangle]
+extern "C" fn get_item(list_ptr: *mut i32, len: usize, index: usize) -> i32 {
+    assert!(index < len, "Index out of bounds");
+    let slice = unsafe { slice::from_raw_parts(list_ptr, len) };
+    let item = slice[index];
+
+    item
+}
+
+#[no_mangle]
+extern "C" fn remove_item(list_ptr: *mut i32, len: usize, index: usize) -> *const i32 {
+    let mut vec = unsafe { Vec::from_raw_parts(list_ptr, len, len) };
+    vec.remove(index);
+    let new_ptr = vec.as_ptr();
+    mem::forget(vec);
+    new_ptr
+}
+
+#[no_mangle]
+extern "C" fn release_list(list_ptr: *mut i32, len: usize) {
+    _ = unsafe { Vec::from_raw_parts(list_ptr, len, len) };
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_get_list() {
+        let mut length: usize = 0;
+        let ptr = new_list(&mut length);
+        let slice = unsafe { std::slice::from_raw_parts(ptr, length) };
+
+        assert_eq!(slice, &[10, 20, 30, 40, 100]);
+        assert_eq!(length, 5);
+    }
+}
